feat: add types for debug/lodash/p-map, return types in ddb.ts
