+++ package.json
@@ -36,8 +36,12 @@
     "@babel/preset-typescript": "7.8.3",
     "@shelf/eslint-config": "0.13.0",
     "@shelf/jest-dynamodb": "1.6.0",
+    "@types/debug": "4.1.5",
     "@types/jest": "24.9.0",
+    "@types/lodash.clonedeep": "4.5.6",
+    "@types/lodash.times": "4.3.6",
     "@types/node": "12",
+    "@types/p-map": "2.0.0",
     "aws-sdk": "2.536.0",
     "eslint": "6.8.0",
     "husky": "4.0.9",
